services/locations/requirements.txt
starlette
uvicorn
asyncpg
python-dotenv
urllib3
aiomysql